refactor(cockpit): extract ProjectDashboardData into sibling module (#1354)
Eighth wedge of the cockpit_ui.py god-module split. Moves the
ProjectDashboardData snapshot container (data-only, __slots__-defined,
~30 attributes) into a new cockpit_project_dashboard_data.py sibling
module and re-exports it from cockpit_ui for back-compat.

- cockpit_ui.py: 18740 -> 18638 LOC (-102)
- in-file diff: 3 inserts / 105 deletions (under 200-line budget)
- new cockpit_project_dashboard_data.py: 139 LOC (incl. module docstring)
- no behavior change; ProjectDashboardData is constructed only in
  cockpit_ui (_gather_project_dashboard) and has zero external deps
  beyond pathlib.Path for type hints
- tests/test_project_dashboard_ui.py keeps its existing
  `from pollypm.cockpit_ui import ProjectDashboardData` import path

Refs #1354.
